tests: factor out some code into conftest
Some definitions in test_manifest.py will be useful in
test_project.py.
